+++ geocoding/lib/geocoding.dart
@@ -1,13 +1,45 @@
 import 'dart:async';
 
-import 'package:flutter/services.dart';
+import 'package:geocoding_platform_interface/geocoding_platform_interface.dart';
 
-class Geocoding {
-  static const MethodChannel _channel =
-      const MethodChannel('geocoding');
+/// Returns a list of [Location] instances found for the supplied address.
+///
+/// In most situations the returned list should only contain one entry.
+/// However in some situations where the supplied address could not be
+/// resolved into a single [Location], multiple [Location] instances may be
+/// returned.
+///
+/// Optionally you can specify a locale in which the results are returned.
+/// When not supplied the currently active locale of the device will be used.
+/// The `localeIdentifier` should be formatted using the syntax:
+/// [languageCode]_[countryCode] (eg. en_US or nl_NL).
+Future<List<Location>> placemarkFromAddress(
+  String address, {
+  String localeIdentifier,
+}) =>
+    GeocodingPlatform.instance.placemarkFromAddress(
+      address,
+      localeIdentifier: localeIdentifier,
+    );
 
-  static Future<String> get platformVersion async {
-    final String version = await _channel.invokeMethod('getPlatformVersion');
-    return version;
-  }
-}
+/// Returns a list of [Placemark] instances found for the supplied
+/// coordinates.
+///
+/// In most situations the returned list should only contain one entry.
+/// However in some situations where the supplied coordinates could not be
+/// resolved into a single [Placemark], multiple [Placemark] instances may be
+/// returned.
+///
+/// Optionally you can specify a locale in which the results are returned.
+/// When not supplied the currently active locale of the device will be used.
+/// The `localeIdentifier` should be formatted using the syntax:
+/// [languageCode]_[countryCode] (eg. en_US or nl_NL).
+Future<List<Placemark>> placemarkFromCoordinates(
+  double latitude,
+  double longitude, {
+  String localeIdentifier,
+}) =>
+    GeocodingPlatform.instance.placemarkFromCoordinates(
+      latitude,
+      longitude,
+    );