+++ geocoding_platform_interface/lib/src/geocoding_platform_interface.dart
@@ -31,18 +31,18 @@ abstract class GeocodingPlatform extends PlatformInterface {
     _instance = instance;
   }
 
-  /// Returns a list of [Placemark] instances found for the supplied address.
+  /// Returns a list of [Location] instances found for the supplied address.
   ///
   /// In most situations the returned list should only contain one entry.
   /// However in some situations where the supplied address could not be
-  /// resolved into a single [Placemark], multiple [Placemark] instances may be
+  /// resolved into a single [Location], multiple [Location] instances may be
   /// returned.
   ///
   /// Optionally you can specify a locale in which the results are returned.
   /// When not supplied the currently active locale of the device will be used.
   /// The `localeIdentifier` should be formatted using the syntax:
   /// [languageCode]_[countryCode] (eg. en_US or nl_NL).
-  Future<List<Placemark>> placemarkFromAddress(
+  Future<List<Location>> placemarkFromAddress(
     String address, {
     String localeIdentifier,
   }) {