+++ geocoding_platform_interface/lib/src/models/models.dart
@@ -1,1 +1,2 @@
+export 'location.dart';
 export 'placemark.dart';