+++ geocoding_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0+2
+
+- Return `location` class from the `placemarkFromAddress` method.
+
+## 1.0.0+1
+
+- Add missing street and toString implementation to Placemark.
+
 ## 1.0.0
 
-- Initial open-source release
+- Initial open-source release.